Time: 33 ms (71.61%), Space: 63.6 MB (89.63%) - LeetHub

+++ 0973-k-closest-points-to-origin/0973-k-closest-points-to-origin.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     vector<vector<int>> kClosest(vector<vector<int>>& points, int k) {
-        priority_queue<pair<long,int>> pq;
+        priority_queue<pair<int,int>> pq;
 
         for(int i=0;i<points.size();i++){
-            long d = (points[i][0]*points[i][0] + points[i][1]*points[i][1]);
+            int d = (points[i][0]*points[i][0] + points[i][1]*points[i][1]);
             pq.push({d,i});
             if(pq.size() > k) pq.pop();
         }
